packages/studyctl/pyproject.toml
[project]
name = "studyctl"
version = "2.1.0"
description = "AuDHD-aware study tool with AI Socratic mentoring, spaced repetition, and content pipeline"
requires-python = ">=3.12"
readme = "README.md"
license = "MIT"
authors = [{name = "Andy Taylor"}]
keywords = ["study", "flashcards", "spaced-repetition", "audhd", "adhd", "socratic", "notebooklm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Education",
    "Topic :: Education",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
content = ["pymupdf>=1.25", "httpx"]
notebooklm = ["notebooklm-py>=0.3.4"]
tui = ["textual>=0.80"]
web = ["fastapi>=0.115", "uvicorn[standard]>=0.34", "httpx>=0.27", "websockets>=12.0"]
mcp = ["mcp[cli]>=1.0.0"]
all = ["studyctl[content,web,notebooklm,tui,mcp]"]

[project.scripts]
studyctl = "studyctl.cli:cli"
studyctl-mcp = "studyctl.mcp.server:main"

[project.urls]
Homepage = "https://github.com/NetDevAutomate/socratic-study-mentor"
Repository = "https://github.com/NetDevAutomate/socratic-study-mentor"
Issues = "https://github.com/NetDevAutomate/socratic-study-mentor/issues"
Documentation = "https://github.com/NetDevAutomate/socratic-study-mentor/tree/main/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/studyctl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--tb=short"
markers = [
    "integration: requires external infrastructure (tmux, real DB, network)",
    "e2e: end-to-end browser tests requiring Playwright + web server",
]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
exclude = ["src/studyctl/tui", "src/studyctl/content", "src/studyctl/cli/_content.py", "src/studyctl/cli/_web.py", "src/studyctl/web", "src/studyctl/mcp"]  # optional deps not in CI

[dependency-groups]
dev = [
    "pytest-playwright>=0.7.2",
]
